requirements.txt
# Core dependencies
mcp>=0.1.0
fastmcp>=0.1.0

# Video processing
opencv-python>=4.9.0
opencv-python-headless>=4.9.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0

# Audio processing
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# LLM integration
ollama>=0.1.0
httpx>=0.25.0

# HTTP server support
uvicorn>=0.24.0
fastapi>=0.104.0
starlette>=0.27.0

# Storage and data
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.0.0
python-json-logger>=2.0.0

# Async support
aiofiles>=23.0.0
asyncio>=3.4.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
tqdm>=4.66.0
python-dateutil>=2.8.0
psutil>=5.9.0